changed UnusedLocalDefinition behavior and deleted UnusedLocalDefinitionNew dependencies.
