+++ org.eclipse.titanium/src/org/eclipse/titanium/markers/utils/RiskFactorCalculator.java
@@ -41,7 +41,6 @@ public class RiskFactorCalculator {
 		CodeSmellType.UNUSED_IMPORT,
 		CodeSmellType.UNUSED_GLOBAL_DEFINITION,
 		CodeSmellType.UNUSED_LOCAL_DEFINITION,
-		CodeSmellType.UNUSED_LOCAL_DEFINITION_NEW,
 		CodeSmellType.UNUSED_FUNTION_RETURN_VALUES,
 		CodeSmellType.UNUSED_STARTED_FUNCTION_RETURN_VALUES,
 		CodeSmellType.INFINITE_LOOP,
