+++ org.eclipse.titanium/src/org/eclipse/titanium/preferences/pages/RiskFactorPreferencePage.java
@@ -53,7 +53,6 @@ public class RiskFactorPreferencePage extends PreferencePage implements IWorkben
 		insertSP(CodeSmellType.UNUSED_IMPORT);
 		insertSP(CodeSmellType.UNUSED_GLOBAL_DEFINITION);
 		insertSP(CodeSmellType.UNUSED_LOCAL_DEFINITION);
-		insertSP(CodeSmellType.UNUSED_LOCAL_DEFINITION_NEW);
 		insertSP(CodeSmellType.UNUSED_FUNTION_RETURN_VALUES);
 		insertSP(CodeSmellType.UNUSED_STARTED_FUNCTION_RETURN_VALUES);
 		insertSP(CodeSmellType.INFINITE_LOOP);
